+++ bitv.ml
@@ -94,6 +94,20 @@ let init n f =
   for i = 0 to pred n do unsafe_set v i (f i) done;
   v
 
+let random n =
+  let v = create n false in
+  let b = v.bits in
+  let n = Bytes.length b in
+  for i = 0 to n / 3 do let j = 3 * i in
+    let bits = Random.bits () in
+    set_byte b j     (bits          land 0xFF);
+    set_byte b (j+1) ((bits lsr  8) land 0xFF);
+    set_byte b (j+2) ((bits lsr 16) land 0xFF)
+  done;
+  for i = 3 * (n / 3) to n - 1 do set_byte b i (Random.int 256) done;
+  normalize v;
+  v
+
 let fill v ofs len b =
   if ofs < 0 || len < 0 || ofs > v.length - len then invalid_arg "Bitv.fill";
   if len > 0 then (
@@ -216,6 +230,19 @@ let pop v =
    However, one has to take care of normalizing the result of [bwnot]
    which introduces ones in highest significant positions. *)
 
+let[@inline always] bw_and_in_place_internal ~dest ~n b1 b2 =
+  for i = 0 to n - 1 do
+    set_byte dest i ((byte b1 i) land (byte b2 i))
+  done
+
+let bw_and_in_place ~dest v1 v2 =
+  let l = v1.length in
+  if l <> v2.length || l <> dest.length then invalid_arg "Bitv.bw_and_in_place";
+  let b1 = v1.bits
+  and b2 = v2.bits in
+  let n = Bytes.length b1 in
+  bw_and_in_place_internal ~dest:dest.bits ~n b1 b2
+
 let bw_and v1 v2 =
   let l = v1.length in
   if l <> v2.length then invalid_arg "Bitv.bw_and";
@@ -223,10 +250,21 @@ let bw_and v1 v2 =
   and b2 = v2.bits in
   let n = Bytes.length b1 in
   let a = Bytes.make n (Char.chr 0) in
-  for i = 0 to n - 1 do
-    set_byte a i ((byte b1 i) land (byte b2 i))
-  done;
+  bw_and_in_place_internal ~dest:a ~n b1 b2;
   { length = l; bits = a }
+
+let[@inline always] bw_or_in_place_internal ~dest ~n b1 b2 =
+  for i = 0 to n - 1 do
+    set_byte dest i ((byte b1 i) lor (byte b2 i))
+  done
+
+let bw_or_in_place ~dest v1 v2 =
+  let l = v1.length in
+  if l <> v2.length || l <> dest.length then invalid_arg "Bitv.bw_or_in_place";
+  let b1 = v1.bits
+  and b2 = v2.bits in
+  let n = Bytes.length b1 in
+  bw_or_in_place_internal ~dest:dest.bits ~n b1 b2
 
 let bw_or v1 v2 =
   let l = v1.length in
@@ -235,33 +273,53 @@ let bw_or v1 v2 =
   and b2 = v2.bits in
   let n = Bytes.length b1 in
   let a = Bytes.make n (Char.chr 0) in
-  for i = 0 to n - 1 do
-    set_byte a i ((byte b1 i) lor (byte b2 i))
-  done;
+  bw_or_in_place_internal ~dest:a ~n b1 b2;
   { length = l; bits = a }
+
+let[@inline always] bw_xor_in_place_internal ~dest ~n b1 b2 =
+  for i = 0 to n - 1 do
+    set_byte dest i ((byte b1 i) lxor (byte b2 i))
+  done
+
+let bw_xor_in_place ~dest v1 v2 =
+  let l = v1.length in
+  if l <> v2.length || l <> dest.length then invalid_arg "Bitv.bw_xor_in_place";
+  let b1 = v1.bits
+  and b2 = v2.bits in
+  let n = Bytes.length b1 in
+  bw_xor_in_place_internal ~dest:dest.bits ~n b1 b2
 
 let bw_xor v1 v2 =
   let l = v1.length in
-  if l <> v2.length then invalid_arg "Bitv.bw_or";
+  if l <> v2.length then invalid_arg "Bitv.bw_xor";
   let b1 = v1.bits
   and b2 = v2.bits in
   let n = Bytes.length b1 in
   let a = Bytes.make n (Char.chr 0) in
-  for i = 0 to n - 1 do
-    set_byte a i ((byte b1 i) lxor (byte b2 i))
-  done;
+  bw_xor_in_place_internal ~dest:a ~n b1 b2;
   { length = l; bits = a }
+
+let[@inline always] bw_not_in_place_internal ~dest ~n b =
+  let a = dest.bits in
+  for i = 0 to n - 1 do
+    set_byte a i (255 land (lnot (byte b i)))
+  done;
+  normalize dest
+
+let bw_not_in_place ~dest v =
+  let l = v.length in
+  if l <> dest.length then invalid_arg "Bitv.bw_not_in_place";
+  let b = v.bits in
+  let n = Bytes.length b in
+  bw_not_in_place_internal ~dest ~n b
 
 let bw_not v =
   let b = v.bits in
   let n = Bytes.length b in
   let a = Bytes.make n (Char.chr 0) in
-  for i = 0 to n - 1 do
-    set_byte a i (255 land (lnot (byte b i)))
-  done;
-  let r = { length = v.length; bits = a } in
-  normalize r;
-  r
+  let dest = { length = v.length; bits = a } in
+  bw_not_in_place_internal ~dest ~n b;
+  dest
 
 (* Coercions to/from lists of integers *)