+++ CHANGES.md
@@ -1,5 +1,7 @@
 
   - fixed negative `max_length` on JavaScript platform
+  - new function `random` (much faster than `init n (fun _ -> Random.bool ())`
+  - new in-place bitwise operations (contribution by Mathieu Barbin)
   - fixed division by zero in rotations of 0-length vectors
     (reported by Nikolaus Huber)
   - fixed integer overflows in bound tests in `fill`, `blit`, `sub`